cpufreq: intel_pstate: Add debugfs file stats
Add stats file in debugfs under driver's parent dir
(pstate_snb) which counts the time in nsecs per requested
P state and the number of times the specific state
was requested.

The file presents the statistics per logical CPU in the
following format. The time is display in msecs:

CPU0
P-state        Time     Count
     16     4882777     23632
     17       21210       174
     18      549781      3300
     19       51171       461
     20       35487       394
     21       18173       219
     22       13752       258
     23        6048       172
     24        7754       177
     25        4587       151
     26        5465       162
     27        1432        47
     28         863        54
     29        1448        50
     30        1030        47
     31        1472        62
     32        2221        68
     33        1869        60
     34        2140        70
     39       85446      3803

...

The file can be used for debugging but also for monitoring
various system workloads.

Also, make the debugfs_parent local as we never remove
the driver's debugfs files.

Signed-off-by: Stratos Karafotis <stratosk@semaphore.gr>